refactor: post_view, contents_detail_page 코드 리팩토링
