+++ lib/write/presentation/screens/contents_detail/contents_detail_page.dart
@@ -1,259 +1,13 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_sns/theme/theme.dart';
-import 'package:flutter_sns/utils/xss.dart';
 import 'package:flutter_sns/write/domain/entities/comments.dart'
     as comment_entity;
-import 'package:flutter_sns/write/domain/entities/posts.dart';
+import 'package:flutter_sns/write/presentation/providers/post_detail_providers.dart';
 import 'package:flutter_sns/write/presentation/screens/contents_detail/widgets/comment_input.dart';
+import 'package:flutter_sns/write/presentation/screens/contents_detail/widgets/comment_section_view.dart';
 import 'package:flutter_sns/write/presentation/screens/contents_detail/widgets/post_contents_view.dart';
 import 'package:flutter_sns/write/presentation/screens/home/widgets/no_glow_scroll_behavior.dart';
-import 'package:flutter_sns/write/presentation/screens/contents_detail/widgets/comment_section_view.dart';
-
-// --- Providers and Helper Functions ---
-
-// Firestore 데이터를 Posts 엔티티로 변환하는 헬퍼 함수
-Posts _postFromFirestore(DocumentSnapshot<Map<String, dynamic>> doc) {
-  final data = doc.data();
-  if (data == null) throw StateError('Missing data for post ${doc.id}');
-
-  // 중첩된 author 객체 처리
-  final authorData = data['author'];
-  final authorMap = (authorData is Map)
-      ? Map<String, dynamic>.from(authorData)
-      : <String, dynamic>{};
-  final author = Author(
-    nickname: authorMap['nickname'] as String? ?? 'Unknown User',
-    profileImageUrl: authorMap['profileImageUrl'] as String?,
-  );
-
-  // 중첩된 stats 객체 처리
-  final statsData = data['stats'];
-  final statsMap = (statsData is Map)
-      ? Map<String, dynamic>.from(statsData)
-      : <String, dynamic>{};
-  final stats = PostStats(
-    likesCount: statsMap['likesCount'] as int? ?? 0,
-    commentsCount: statsMap['commentsCount'] as int? ?? 0,
-  );
-
-  // 이미지 리스트 처리
-  final imagesData = data['images'];
-  final images = (imagesData is List)
-      ? imagesData.map((item) => item.toString()).toList()
-      : <String>[];
-
-  // 타임스탬프 처리
-  final createdAtData = data['createdAt'];
-  final updatedAtData = data['updatedAt'];
-  final createdAt = (createdAtData is Timestamp)
-      ? createdAtData.toDate()
-      : DateTime.now();
-  final updatedAt = (updatedAtData is Timestamp)
-      ? updatedAtData.toDate()
-      : DateTime.now();
-
-  return Posts(
-    id: doc.id,
-    authorId: data['authorId'] as String? ?? '',
-    author: author,
-    category: data['category'] as String? ?? '',
-    mode: data['mode'] as String? ?? '',
-    title: data['title'] as String? ?? '',
-    content: data['content'] as String? ?? '',
-    images: images,
-    stats: stats,
-    createdAt: createdAt,
-    updatedAt: updatedAt,
-    reportCount: data['reportCount'] as int? ?? 0,
-  );
-}
-
-// Firestore 데이터를 Comments 엔티티로 변환하는 헬퍼 함수
-comment_entity.Comments _commentFromFirestore(
-  DocumentSnapshot<Map<String, dynamic>> doc,
-) {
-  final data = doc.data();
-  if (data == null) throw StateError('Missing data for comment ${doc.id}');
-
-  // 중첩된 author 객체 처리 (더 안전하게)
-  final authorData = data['author'];
-  final authorMap = (authorData is Map)
-      ? Map<String, dynamic>.from(authorData)
-      : <String, dynamic>{};
-  final author = comment_entity.Author(
-    nickname: authorMap['nickname'] as String? ?? 'Unknown User',
-    profileImageUrl: authorMap['profileImageUrl'] as String?,
-  );
-
-  // 타임스탬프 처리
-  final createdAtData = data['createdAt'];
-  final updatedAtData = data['updatedAt'];
-  final createdAt = (createdAtData is Timestamp)
-      ? createdAtData.toDate()
-      : DateTime.now();
-  final updatedAt = (updatedAtData is Timestamp)
-      ? updatedAtData.toDate()
-      : DateTime.now();
-
-  // reportCount 처리
-  final reportCountData = data['reportCount'];
-  final reportCount = (reportCountData is num) ? reportCountData.toInt() : 0;
-
-  return comment_entity.Comments(
-    id: doc.id,
-    postId: data['postId'] as String? ?? '',
-    authorId: data['userId'] as String? ?? '',
-    author: author,
-    content: data['content'] as String? ?? '',
-    createdAt: createdAt,
-    updatedAt: updatedAt,
-    reportCount: reportCount,
-  );
-}
-
-// 특정 게시물 하나의 스트림을 제공하는 프로바이더
-final postProvider = StreamProvider.family<Posts, String>((ref, postId) {
-  return FirebaseFirestore.instance
-      .collection('posts')
-      .doc(postId)
-      .snapshots()
-      .map(_postFromFirestore);
-});
-
-// 특정 게시물의 댓글 목록 스트림을 제공하는 프로바이더
-final commentsProvider =
-    StreamProvider.family<List<comment_entity.Comments>, String>((ref, postId) {
-      return FirebaseFirestore.instance
-          .collection('comments')
-          .where('postId', isEqualTo: postId)
-          .orderBy('createdAt', descending: false)
-          .snapshots()
-          .map((snapshot) => snapshot.docs.map(_commentFromFirestore).toList());
-    });
-
-// --- Service and Provider for adding comments ---
-
-class CommentService {
-  final FirebaseFirestore _firestore;
-  final FirebaseAuth _auth;
-
-  CommentService(this._firestore, this._auth);
-
-  Future<void> addComment({
-    required String postId,
-    required String content,
-  }) async {
-    final currentUser = _auth.currentUser;
-    if (currentUser == null) {
-      throw Exception('로그인이 필요합니다.');
-    }
-
-    final sanitizedContent = XssFilter.sanitize(content);
-    if (sanitizedContent.isEmpty) {
-      throw Exception('댓글 내용을 입력해주세요.');
-    }
-
-    // Firestore에서 현재 사용자의 닉네임과 프로필 이미지 URL 가져오기
-    final userDoc = await _firestore
-        .collection('users')
-        .doc(currentUser.uid)
-        .get();
-    final userData = userDoc.data();
-    final nickname = userData?['nickname'] as String? ?? '이름없음';
-    final profileImageUrl = userData?['profileImageUrl'] as String?;
-
-    final newCommentRef = _firestore.collection('comments').doc();
-    // final postRef = _firestore.collection('posts').doc(postId);
-
-    final commentData = {
-      'postId': postId,
-      'userId': currentUser.uid,
-      'author': {'nickname': nickname, 'profileImageUrl': profileImageUrl},
-      'content': sanitizedContent,
-      'createdAt': FieldValue.serverTimestamp(),
-      'updatedAt': FieldValue.serverTimestamp(),
-      'reportCount': 0,
-    };
-
-    // Firestore 보안 규칙에 따라, 다른 사람의 게시물('posts' 문서)을 수정할 권한이 없습니다.
-    // 따라서 댓글을 추가할 때 게시물의 댓글 수를 직접 업데이트하는 로직을 제거하고,
-    // 'comments' 컬렉션에 새 댓글을 추가하는 작업만 수행합니다.
-    // 게시물의 댓글 수를 실시간으로 정확하게 반영하려면 서버 측 로직(예: Cloud Function)을 사용해야 합니다.
-    await newCommentRef.set(commentData);
-  }
-}
-
-final commentServiceProvider = Provider((ref) {
-  return CommentService(FirebaseFirestore.instance, FirebaseAuth.instance);
-});
-
-// --- Service and Provider for post interactions (likes) ---
-
-class PostInteractionService {
-  final FirebaseFirestore _firestore;
-  final FirebaseAuth _auth;
-
-  PostInteractionService(this._firestore, this._auth);
-
-  Future<void> toggleLike({required String postId}) async {
-    final currentUser = _auth.currentUser;
-    if (currentUser == null) {
-      throw Exception('로그인이 필요합니다.');
-    }
-    final userId = currentUser.uid;
-    // 규칙에 따라 루트 'likes' 컬렉션을 사용하고, 문서 ID를 조합하여 만듭니다.
-    final likeId = '${postId}_$userId';
-    final likeRef = _firestore.collection('likes').doc(likeId);
-    final likeDoc = await likeRef.get();
-
-    if (likeDoc.exists) {
-      // Unlike the post
-      await likeRef.delete();
-    } else {
-      // Like the post. 규칙에 맞게 postId와 userId를 저장합니다.
-      await likeRef.set({
-        'postId': postId,
-        'userId': userId,
-        'createdAt': FieldValue.serverTimestamp(),
-      });
-    }
-  }
-}
-
-final postInteractionServiceProvider = Provider((ref) {
-  return PostInteractionService(
-    FirebaseFirestore.instance,
-    FirebaseAuth.instance,
-  );
-});
-
-final postLikesCountProvider = StreamProvider.family<int, String>((
-  ref,
-  postId,
-) {
-  return FirebaseFirestore.instance
-      .collection('likes')
-      .where('postId', isEqualTo: postId)
-      .snapshots()
-      .map((snapshot) => snapshot.size);
-});
-
-final isPostLikedProvider = StreamProvider.family<bool, String>((ref, postId) {
-  final userId = FirebaseAuth.instance.currentUser?.uid;
-  if (userId == null) {
-    return Stream.value(false);
-  }
-  // 루트 'likes' 컬렉션에서 조합된 ID로 문서를 직접 조회합니다.
-  final likeId = '${postId}_$userId';
-  return FirebaseFirestore.instance
-      .collection('likes')
-      .doc(likeId)
-      .snapshots()
-      .map((snapshot) => snapshot.exists);
-});
 
 class ContentsDetailPage extends ConsumerWidget {
   const ContentsDetailPage({super.key, required this.postId});
@@ -295,11 +49,8 @@ class ContentsDetailPage extends ConsumerWidget {
           behavior: NoGlowScrollBehavior(),
           child: ListView(
             children: [
-              // 1. 게시물 본문 위젯
-              PostContentView(
-                post: post,
-                commentCount: commentsAsyncValue.asData?.value.length,
-              ),
+              // 1. 게시물 본문 위젯 (댓글 수는 PostActions 위젯 내부에서 실시간으로 가져옵니다)
+              PostContentView(post: post),
               // 2. 구분선
               const Divider(height: 1, thickness: 1, color: AppColors.n300),
               // 3. 댓글 목록 위젯
