+++ lib/write/presentation/screens/contents_detail/widgets/post_contents_view.dart
@@ -8,8 +8,7 @@ import 'package:intl/intl.dart';
 /// 게시물의 제목, 내용, 이미지, 상호작용 버튼 등을 포함하는 위젯
 class PostContentView extends StatelessWidget {
   final Posts post;
-  final int? commentCount;
-  const PostContentView({super.key, required this.post, this.commentCount});
+  const PostContentView({super.key, required this.post});
 
   /// DateTime 객체를 '몇 분 전'과 같은 상대 시간 문자열로 변환
   String _formatTimeAgo(DateTime dateTime) {
@@ -99,11 +98,7 @@ class PostContentView extends StatelessWidget {
           const SizedBox(height: 16),
 
           // 좋아요 및 댓글 수
-          PostActions(
-            postId: post.id,
-            likeCount: post.stats.likesCount,
-            commentCount: commentCount ?? post.stats.commentsCount,
-          ),
+          PostActions(postId: post.id, likeCount: post.stats.likesCount),
         ],
       ),
     );
