+++ lib/write/presentation/screens/home/home_page.dart
@@ -2,68 +2,10 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_sns/write/domain/entities/posts.dart';
+import 'package:flutter_sns/write/presentation/screens/contents_detail/widgets/firestore_mapper.dart';
 import 'package:flutter_sns/write/presentation/screens/home/widgets/no_glow_scroll_behavior.dart';
 import 'package:flutter_sns/write/presentation/screens/home/widgets/post_view.dart';
 import 'package:flutter_sns/write/presentation/screens/home/widgets/top_tab_bar.dart';
-
-// Firestore 데이터를 Posts 엔티티로 변환하는 헬퍼 함수
-Posts _postFromFirestore(DocumentSnapshot<Map<String, dynamic>> doc) {
-  final data = doc.data();
-  if (data == null) {
-    throw StateError('Missing data for post ${doc.id}');
-  }
-
-  // 중첩된 author 객체 처리
-  final authorData = data['author'];
-  final authorMap = (authorData is Map)
-      ? Map<String, dynamic>.from(authorData)
-      : <String, dynamic>{};
-  final author = Author(
-    nickname: authorMap['nickname'] as String? ?? 'Unknown User',
-    profileImageUrl: authorMap['profileImageUrl'] as String?,
-  );
-
-  // 중첩된 stats 객체 처리
-  final statsData = data['stats'];
-  final statsMap = (statsData is Map)
-      ? Map<String, dynamic>.from(statsData)
-      : <String, dynamic>{};
-  final stats = PostStats(
-    likesCount: statsMap['likesCount'] as int? ?? 0,
-    commentsCount: statsMap['commentsCount'] as int? ?? 0,
-  );
-
-  // 이미지 리스트 처리
-  final imagesData = data['images'];
-  final images = (imagesData is List)
-      ? imagesData.map((item) => item.toString()).toList()
-      : <String>[];
-
-  // 타임스탬프 처리
-  final createdAtData = data['createdAt'];
-  final updatedAtData = data['updatedAt'];
-  final createdAt = (createdAtData is Timestamp)
-      ? createdAtData.toDate()
-      : DateTime.now();
-  final updatedAt = (updatedAtData is Timestamp)
-      ? updatedAtData.toDate()
-      : DateTime.now();
-
-  return Posts(
-    id: doc.id,
-    authorId: data['authorId'] as String? ?? '',
-    author: author,
-    category: data['category'] as String? ?? '',
-    mode: data['mode'] as String? ?? '',
-    title: data['title'] as String? ?? '',
-    content: data['content'] as String? ?? '',
-    images: images,
-    stats: stats,
-    createdAt: createdAt,
-    updatedAt: updatedAt,
-    reportCount: data['reportCount'] as int? ?? 0,
-  );
-}
 
 // 게시물 목록 스트림을 제공하는 프로바이더
 final postsStreamProvider = StreamProvider<List<Posts>>((ref) {
@@ -73,7 +15,7 @@ final postsStreamProvider = StreamProvider<List<Posts>>((ref) {
       .orderBy('createdAt', descending: true)
       .snapshots()
       .map((snapshot) {
-        return snapshot.docs.map((doc) => _postFromFirestore(doc)).toList();
+        return snapshot.docs.map((doc) => postFromFirestore(doc)).toList();
       });
 });
